requirements.txt
langchain-groq
langchain-community
langchain-tavily
langgraph
python-dotenv
google-generativeai
tavily-python
fastapi
uvicorn[standard]
python-multipart
